1.19

+++ module-compat/alloy-forgery/src/main/java/lol/bai/megane/module/alloyforgery/mixin/AccessorForgeControllerBlockEntity.java
@@ -12,7 +12,7 @@ public interface AccessorForgeControllerBlockEntity {
     ForgeDefinition getForgeDefinition();
 
     @Accessor
-    int getFuel();
+    float getFuel();
 
 
 }
