+++ module-compat/alloy-forgery/src/main/java/lol/bai/megane/module/alloyforgery/provider/ForgeControllerEnergyProvider.java
@@ -15,7 +15,7 @@ public class ForgeControllerEnergyProvider extends EnergyProvider<ForgeControlle
 
     @Override
     public long getStored() {
-        return access.getFuel();
+        return (long) access.getFuel();
     }
 
     @Override
